+++ Portal/Model/Server/IServerRepository.cs
@@ -5,9 +5,9 @@ namespace OpenSim.Portal.Model
     public interface IServerRepository
     {
         IQueryable<Server> GetAll();
-        Server Get(long id);
+        Server Get(int id);
         void Add(Server server);
-        Server Remove(long id);
+        Server Remove(int id);
         void Update(Server server);
     }
 }
